+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 ## Change log
 ----------------------
 
+Version 2.22
+-------------
+
+CHANGED:
+
+- update of parent version to 4
+- update of jcommons-lang to 4.34
+- update of test-objects version to 4.28
+- update of silly-collections version to 4.31
+- update of jobject-extensions to 1.11
+- update of meanbean-factories to 1.1
+
 Version 2.21
 -------------
 
@@ -9,6 +21,7 @@ CHANGED:
 - update of parent version to 3.11
 - removed unneeded .0 at the end of version
 - update of file-worker version to 4.20
+- update of silly-collections version to 4.28.0
 
 Version 2.20.0
 -------------